# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /compilations:
    get:
      operationId: listCompilations
      tags: [screener]
      summary: List my research compilations
      responses:
        "200":
          description: Compilations
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Compilation" }
    post:
      operationId: createCompilation
      tags: [screener]
      summary: Group selected saved screens
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CompilationInput" }
      responses:
        "201":
          description: Saved compilation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Compilation" }
  /compilations/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    get:
      operationId: getCompilation
      tags: [screener]
      summary: Retrieve compilation and included screens
      responses:
        "200":
          description: Compilation detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CompilationDetail" }
    delete:
      operationId: deleteCompilation
      tags: [screener]
      summary: Delete a compilation
      responses:
        "204": { description: Deleted }
  /explanations:
    post:
      operationId: createExplanation
      tags: [screener]
      summary: Ask an evidence-grounded question about a saved screen
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ExplanationInput" }
      responses:
        "200":
          description: Cited answer
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Explanation" }
  /screens/{id}/explanations:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    get:
      operationId: listExplanations
      tags: [screener]
      summary: My AI questions for a saved screen
      responses:
        "200":
          description: Question history
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Explanation" }
...
  /billing/checkout:
    post:
      operationId: createCheckout
      tags: [screener]
      summary: Begin a test-mode or approved live subscription upgrade
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CheckoutInput" }
      responses:
        "200":
          description: Hosted checkout
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BillingLink" }
...
components:
  schemas:
...
    TimePoint:
      type: object
      required: [month, visits]
      properties:
        month: { type: string }
        visits: { type: number }
    ShareItem:
      type: object
      required: [label, share]
      properties:
        label: { type: string }
        share: { type: number }
    DomainProfile:
      type: object
      required: [domain, monthlyVisits, averageVisits, trafficChangePercent, channels, geography, topCountry, topChannel]
      properties:
        domain: { type: string }
        monthlyVisits:
          type: array
          items: { $ref: "#/components/schemas/TimePoint" }
        averageVisits: { type: number }
        trafficChangePercent: { type: ["number", "null"] }
        channels:
          type: array
          items: { $ref: "#/components/schemas/ShareItem" }
        geography:
          type: array
          items: { $ref: "#/components/schemas/ShareItem" }
        topCountry:
          type: ["object", "null"]
          properties:
            label: { type: string }
            share: { type: number }
        topChannel:
          type: ["object", "null"]
          properties:
            label: { type: string }
            share: { type: number }
    Finding:
      type: object
      required: [id, severity, title, metric, value, period, source, sourceUrl, whyItMatters, verificationQuestion]
      properties:
        id: { type: string }
        severity: { type: string, enum: [watch, context] }
        title: { type: string }
        metric: { type: string }
        value: { type: string }
        period: { type: string }
        source: { type: string }
        sourceUrl: { type: string }
        whyItMatters: { type: string }
        verificationQuestion: { type: string }
    ScreenReport:
      type: object
      required: [generatedAt, period, sourceNotice, target, comparison, findings, summary, sellerQuestions, limitations]
      properties:
        generatedAt: { type: string }
        id: { type: string }
        period: { type: string }
        sourceNotice: { type: string }
        target: { $ref: "#/components/schemas/DomainProfile" }
        comparison: { $ref: "#/components/schemas/DomainProfile" }
        additionalProfiles:
          type: array
          items: { $ref: "#/components/schemas/DomainProfile" }
        findings:
          type: array
          items: { $ref: "#/components/schemas/Finding" }
        summary: { type: string }
        sellerQuestions:
          type: array
          items: { type: string }
        limitations:
          type: array
          items: { type: string }
    SavedScreen:
      type: object
      required: [id, createdAt, targetDomain, comparisonDomains, report]
      properties:
        id: { type: string }
        createdAt: { type: string }
        targetDomain: { type: string }
        comparisonDomains:
          type: array
          items: { type: string }
        report: { $ref: "#/components/schemas/ScreenReport" }
...
    CompilationInput:
      type: object
      required: [title, screenIds]
      properties:
        title: { type: string, minLength: 1, maxLength: 120 }
        screenIds:
          type: array
          minItems: 1
          maxItems: 100
          items: { type: string }
    Compilation:
      type: object
      required: [id, title, screenIds, summary, createdAt]
      properties:
        id: { type: string }
        title: { type: string }
        screenIds:
          type: array
          items: { type: string }
        summary: { type: string }
        createdAt: { type: string }
    CompilationDetail:
      type: object
      required: [id, title, screenIds, summary, createdAt, screens]
      properties:
        id: { type: string }
        title: { type: string }
        screenIds:
          type: array
          items: { type: string }
        summary: { type: string }
        createdAt: { type: string }
        screens:
          type: array
          items: { $ref: "#/components/schemas/SavedScreen" }
    ExplanationInput:
      type: object
      required: [screenId, question, requestId]
      properties:
        screenId: { type: string }
        question: { type: string, minLength: 5, maxLength: 800 }
        requestId: { type: string, minLength: 8, maxLength: 100 }
    Citation:
      type: object
      required: [label, url, period, retrievedAt]
      properties:
        label: { type: string }
        url: { type: string }
        period: { type: string }
        retrievedAt: { type: string }
    Explanation:
      type: object
      required: [id, screenId, question, answer, citations, createdAt]
      properties:
        id: { type: string }
        screenId: { type: string }
        question: { type: string }
        answer: { type: string }
        citations:
          type: array
          items: { $ref: "#/components/schemas/Citation" }
        createdAt: { type: string }
...
    CheckoutInput:
      type: object
      required: [tier]
      properties:
        tier: { type: string, enum: [pro, team, enterprise] }
    BillingLink:
      type: object
      required: [url]
      properties:
        url: { type: string }