# lib/api-zod/src/generated/types/compilation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Compilation {
  id: string;
  title: string;
  screenIds: string[];
  summary: string;
  createdAt: string;
}