# lib/api-zod/src/generated/types/explanationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ExplanationInput {
  screenId: string;
  /**
     * @minLength 5
     * @maxLength 800
     */
  question: string;
  /**
     * @minLength 8
     * @maxLength 100
     */
  requestId: string;
}